move statistics to its own file
