+++ components/Countdown/countdown.module.scss
@@ -1,5 +1,4 @@
 .countdownWrapper {
-  padding: 0.25rem;
   padding-left: 0;
 }
 
@@ -9,8 +8,4 @@ div .countdownPrefix {
 
 .time {
   color: rgb(117, 193, 255);
-}
-
-.other {
-  font-weight: bold;
 }
